add-event

+++ src/main/java/org/mum/wap/dao/EventDao.java
@@ -83,7 +83,7 @@ public class EventDao {
         return lstEvents;
     }
 
-    public static long addEvent(String title, String description, LocalDateTime startDateTime, int status, String currentLocation, int created_by) {
+    public static long addEvent(String title, String description, LocalDateTime startDateTime, int status, String currentLocation, long created_by) {
 
         Helper helper = new Helper();
         helper.makeJDBCConnection();
